# docs/api/admin-openapi.yaml
openapi: 3.0.3
info:
  title: IoLink Admin API — TARGET
  version: 1.0.0-planned
  description: TARGET / 待实现、待验收：本文件保留 M0–M5 管理 v1 契约，并包含 M6a products/models/assignment 增量；并非现有实现已通过验收的声明。原有 v1 操作路径保持不变；新增目标操作见各 summary。所有 JSON 字段 snake_case；集合始终返回数组，无结果为 []，不得返回 null。时间统一返回 UTC RFC3339
    字符串。所有参数越界/非法值返回 400，不静默回退。所有受保护操作必须逐请求校验身份及资源归属。管理员使用aid claim和管理签名密钥，小程序使用uid claim和用户签名密钥；均由IOLINK_SECRET_KEY以不同标签派生（逻辑密钥，不是新增环境变量），不得互验。管理权限覆盖本实例全部业务资源，管理员不会自动成为
    farm.owner_id。管理集合均支持 limit/offset，默认按 id 升序稳定排序；报警按 created_at DESC,id DESC。新增资源 POST 非幂等，网络结果未知时须查询核实，不能盲目重试；PUT 为完整字段替换且重复执行结果相同；确认操作及 DELETE 幂等，重复成功不产生新副作用。
servers:
- url: http://localhost:8080/admin/v1
...
security:
- BearerAuth: []
paths:
...
  /ponds:
    get:
      tags:
      - ponds
      summary: 池塘列表及最新状态
      description: 按目标契约实现并以集成测试验收。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pond'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: farm_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
      - *id001
      - *id002
    post:
      tags:
      - ponds
      summary: 新建池塘
      description: farm_id 不存在返回 404；成功初始 status=normal、device_count=0、latest=null。POST 非幂等。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pond'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PondReq'
  /ponds/{id}:
    put:
      tags:
      - ponds
      summary: 修改池塘
      description: 幂等；farm_id 必须保持不变。省略 area_mu 置 0。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PondUpdateReq'
    delete:
      tags:
      - ponds
      summary: 删除池塘
      description: 无记录或已删除返回204；有任何设备记录（含停用）、规则、历史遥测或报警引用返回409。M0–M5不提供绕过该限制的归档/硬删入口，不级联删除历史。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
  /devices:
    get:
      tags:
      - devices
      summary: 设备列表
      description: 默认返回未停用设备，include_disabled=true 包含停用设备；不返回明文 secret 或摘要。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: pond_id
        in: query
        required: false
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: include_disabled
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - *id001
      - *id002
    post:
      tags:
      - devices
      summary: 注册设备并返回一次性 secret
      description: 强制绑定已存在 pond_id；无裸设备/裸解绑状态。device_no 服务端生成且永久唯一。secret 只在本次响应显示，服务端仅存摘要。POST 非幂等，响应丢失不得自动重试；管理员核实后停用旧记录并重新注册。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                required:
                - device_no
                - secret
                - pond_id
                - name
                - model
                - status
                - report_interval
                properties:
                  device_no:
                    type: string
                  secret:
                    type: string
                    minLength: 64
                    maxLength: 64
                    pattern: ^[0-9a-f]{64}$
                    description: 32字节随机值的64字符hex编码，仅本次返回。
                  pond_id: &id004
                    type: integer
                    format: int64
                    minimum: 1
                  name:
                    type: string
                  model:
                    type: string
                  status:
                    type: string
                    enum:
                    - offline
                  report_interval:
                    type: integer
                    enum:
                    - 60
                    - 300
                additionalProperties: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - pond_id
              properties:
                pond_id: *id004
                name:
                  type: string
                  default: ''
                model:
                  type: string
                  default: ''
                report_interval:
                  type: integer
                  enum:
                  - 60
                  - 300
                  description: 缺省继承服务器当时IOLINK_REPORT_INTERVAL默认值（60秒），写入设备记录；固件需同周期。
              additionalProperties: false
  /devices/{device_no}:
    get:
      tags:
      - devices
      summary: '[新增目标] 设备详情及最新数据'
      description: 管理端可读取停用设备详情及已保留影子；latest 只来自设备当前池塘的样本，没有样本为 null。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - &id005
        name: device_no
        in: path
        required: true
        schema:
          type: string
          minLength: 1
    delete:
      tags:
      - devices
      summary: 停用设备并保留历史
      description: 幂等；软删除保留历史遥测、报警和池塘绑定；撤销 MQTT 连接及重连权限，立即使其退出在线计数，不再接受新消息。重复停用或不存在均 204。device_no 不得复用；管理历史报警仍可查看和确认。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id005
...
  /alarms:
    get:
      tags:
      - alarms
      summary: 报警列表（管理视角）
      description: 按目标契约实现并以集成测试验收。
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alarm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - name: level
        in: query
        required: false
        schema:
          type: string
          enum:
          - critical
          - warning
      - name: only_unconfirmed
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - *id001
      - *id002
  /alarms/{id}/confirm:
    post:
      tags:
      - alarms
      summary: 确认报警
      description: 幂等：已确认仍 204，保留首次 confirmed_at。确认不删除报警或历史样本；不存在返回 404。
      responses:
        '204':
          description: 成功；响应无正文
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
      parameters:
      - *id003
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数、JSON 或业务校验失败；error=invalid_request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 凭据缺失/过期/错误、错误 token 类型或权限主体失效；error=unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: 资源不存在或当前主体无权访问；统一 error=not_found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 关联依赖阻止删除等状态冲突；error=conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: 内部故障；error=internal_error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          minLength: 1
          description: 稳定错误码；不得返回数据库、口令、密钥或内部堆栈。
      additionalProperties: false
...
    WaterLatest:
      type: object
      required:
      - device_no
      - pond_id
      - ts
      - temperature
      - dissolved_oxygen
      - ph
      - turbidity
      - salinity
      - signal
      - timestamps
      - report_interval
      properties:
        device_no:
          type: string
        pond_id: *id006
        ts: &id007
          type: string
          format: date-time
        temperature:
          type: number
          nullable: true
        dissolved_oxygen:
          type: number
          nullable: true
        ph:
          type: number
          nullable: true
        turbidity:
          type: number
          nullable: true
        salinity:
          type: number
          nullable: true
        signal:
          type: integer
          nullable: true
        timestamps:
          type: object
          required:
          - temperature
          - dissolved_oxygen
          - ph
          - turbidity
          - salinity
          - signal
          properties:
            temperature:
              type: string
              format: date-time
              nullable: true
            dissolved_oxygen:
              type: string
              format: date-time
              nullable: true
            ph:
              type: string
              format: date-time
              nullable: true
            turbidity:
              type: string
              format: date-time
              nullable: true
            salinity:
              type: string
              format: date-time
              nullable: true
            signal:
              type: string
              format: date-time
              nullable: true
          additionalProperties: false
        report_interval:
          type: integer
          enum:
          - 60
          - 300
          default: 60
          description: 秒，所选设备当前有效上报间隔
      additionalProperties: false
      nullable: true
      description: 无可见有效属性时为 null；有影子时所有键存在，尚未采集指标及对应 timestamps 项为 null。按属性合并最新已知值，未上报或 JSON null 不清空旧值；ts 是最近一次有效属性更新时间，各属性时间见 timestamps。UI 按 now - timestamps[metric]
        > 3×report_interval 标记字段过期。pond_id 为该影子当前所属塘；signal 单位 dBm。
    Device:
      type: object
      required:
      - id
      - device_no
      - pond_id
      - name
      - model
      - status
      - last_seen_at
      - created_at
      - disabled_at
      - report_interval
      properties:
        id: *id006
        device_no: &id008
          type: string
        pond_id: *id006
        name: &id009
          type: string
        model: &id010
          type: string
        status: &id011
          type: string
          enum:
          - online
          - offline
        last_seen_at: &id012
          type: string
          format: date-time
          nullable: true
        created_at: *id007
        disabled_at: &id013
          type: string
          format: date-time
          nullable: true
        report_interval:
          type: integer
          enum:
          - 60
          - 300
          description: 注册确定的预期上报周期，秒。
      additionalProperties: false
      description: 小程序只返回未停用设备，管理端可使用 include_disabled 查询停用设备。name/model 未设置时为空字符串；永不返回 secret 或摘要。
    Alarm:
      type: object
      required:
      - id
      - device_no
      - pond_id
      - metric
      - current_value
      - threshold
      - level
      - message
      - confirmed_at
      - created_at
      properties:
        id: *id006
        device_no:
          type: string
        pond_id: *id006
        metric:
          type: string
          enum:
          - temperature
          - dissolved_oxygen
          - ph
          - turbidity
          - salinity
        current_value:
          type: number
        threshold:
          type: number
        level:
          type: string
          enum:
          - critical
          - warning
        message:
          type: string
        confirmed_at:
          type: string
          format: date-time
          nullable: true
        created_at: *id007
      additionalProperties: false
      description: pond_id 为触发时快照；level 来自命中的规则配置，不能按上/下阈值自动映射。
    DeviceDetail:
      type: object
      required:
      - id
      - device_no
      - pond_id
      - name
      - model
      - status
      - last_seen_at
      - created_at
      - disabled_at
      - latest
      - report_interval
      properties:
        id: *id006
        device_no: *id008
        pond_id: *id006
        name: *id009
        model: *id010
        status: *id011
        last_seen_at: *id012
        created_at: *id007
        disabled_at: *id013
        latest:
          $ref: '#/components/schemas/WaterLatest'
        report_interval:
          type: integer
          enum:
          - 60
          - 300
          description: 注册确定的预期上报周期，秒。
      additionalProperties: false
...
    Pond:
      type: object
      required:
      - id
      - farm_id
      - name
      - area_mu
      - created_at
      - status
      - device_count
      - latest
      properties:
        id: *id004
        farm_id: *id004
        name:
          type: string
        area_mu:
          type: number
          minimum: 0
        created_at: *id016
        status:
          type: string
          enum:
          - normal
          - warning
          - critical
          description: 所有未确认报警最高级别 critical > warning > normal；不得先截断报警列表。
        device_count:
          type: integer
          minimum: 0
        latest:
          $ref: '#/components/schemas/WaterLatest'
      additionalProperties: false
    PondReq:
      type: object
      required:
      - farm_id
      - name
      properties:
        farm_id: *id004
        name:
          type: string
          minLength: 1
        area_mu:
          type: number
          minimum: 0
          default: 0
      additionalProperties: false
    PondUpdateReq:
      type: object
      required:
      - farm_id
      - name
      properties:
        farm_id:
          type: integer
          format: int64
          minimum: 1
          description: 必须与原值一致；M0–M5 不支持池塘跨养殖场迁移，变化返回 400。
        name:
          type: string
          minLength: 1
        area_mu:
          type: number
          minimum: 0
          default: 0
      additionalProperties: false
...